build(clients): upgrade @qawolf/api-contracts to 0.53.0
